+++ src/Exception/WriteOperationException.php
@@ -1,0 +1,13 @@
+<?php
+declare(strict_types=1);
+
+namespace Kununu\Elasticsearch\Exception;
+
+/**
+ * Class WriteException
+ *
+ * @package Kununu\Elasticsearch\Exception
+ */
+class WriteOperationException extends RepositoryException
+{
+}